requirements.txt
flask==1.1.2
papermill==2.1.3
pretty-cron==1.2.0
APScheduler==3.6.3
gevent==20.6.2
pycron==3.0.0
pandas==1.1.2
daemonize==2.5.0
escapism==1.0.1
notebook==6.1.4
ipython==7.18.1
pretty_cron==1.2.0
ipykernel==5.3.4
requests==2.24.0
Flask-Classful==0.14.2
sentry-sdk[flask]==0.17.6
pytest-xdist==2.1.0
pytest-flask==1.0.0